requirements.txt
streamlit
mistralai
langchain
faiss     
pypdf2      
python-docx    
pandas        
openpyxl         
python-dotenv
requests      
streamlit-feedback
sqlalchemy
plotly